planet_imagery_browser_app/requirements.txt
# Planet Imagery Browser - Python Dependencies

# Planet SDK
planet>=2.0.0

# Image processing
Pillow>=9.0.0

# HTTP requests
requests>=2.28.0

# Timezone handling
pytz>=2023.3

# Web framework (for Streamlit version)
streamlit>=1.28.0
pandas>=1.5.0

# Desktop GUI (for Tkinter version - usually included with Python)
# tkinter - comes with Python standard library

# Interactive map preview (embedded Chromium browser)
cefpython3>=66.0

# Optional: GDAL for GeoTIFF support
# Install separately with: conda install -c conda-forge gdal
# gdal>=3.4.0

